+++ src/datagrid/datagrid.reducer.js
@@ -75,6 +75,11 @@ export default function datagridReducer(state = INITIAL_STATE, action) {
           'createData',
         ], List(), items => items.push(Immutable.fromJS(action.columnDefaultValues)));
 
+    case TYPES.PLATFORM_DATAGRID_REMOVE_ITEM:
+      return state
+        .deleteIn([action.id, 'data', action.index])
+        .deleteIn([action.id, 'allData', action.index]);
+
     case TYPES.PLATFORM_DATAGRID_REMOVE_NEW_ITEM:
       return state.deleteIn([action.id, 'createData', action.index]);